re-consolidate
